backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pandas==2.2.0
pydantic==2.5.3
python-dotenv==1.0.0
langchain==0.1.0
langchain-openai==0.0.5
openai==1.10.0
sentence-transformers==2.2.2
python-multipart==0.0.6
openpyxl==3.1.2
scikit-learn==1.4.0
numpy==1.26.3